requirements.txt
# Flask
flask
werkzeug
pyopenssl==24.1.0
requests
# Server
gunicorn==22.0.0
# Additional Requirements
## Add your additional requirements here
authlib==1.1.0
sentry-sdk[Flask]

# Core
pydantic>=2.0.0
PyYAML>=6.0.1


#firecrawl
firecrawl-py

# WTForms
Flask-WTF
